+++ src/operators/test/mechanics/testNonlinearMechanics_apply.cpp
@@ -58,13 +58,13 @@ static void myTest( AMP::UnitTest *ut, const std::string &exeName )
 
     testNonlinOperator->apply( solVec, resVec );
 
-    double resNorm1 = resVec->L2Norm();
+    double resNorm1 = static_cast<double>( resVec->L2Norm() );
 
     AMP::pout << "resNorm1 = " << resNorm1 << std::endl;
 
     testNonlinOperator->apply( solVec, resVec );
 
-    double resNorm2 = resVec->L2Norm();
+    double resNorm2 = static_cast<double>( resVec->L2Norm() );
 
     AMP::pout << "resNorm2 = " << resNorm2 << std::endl;
