+++ src/solvers/test/mechanics/testLinearElasticity_patch_1.cpp
@@ -76,17 +76,17 @@ static void linearElasticTest( AMP::UnitTest *ut, const std::string &exeName )
 
         dirichletVecOp->apply( nullVec, mechRhsVec );
 
-        double rhsNorm = mechRhsVec->L2Norm();
+        double rhsNorm = static_cast<double>( mechRhsVec->L2Norm() );
 
         AMP::pout << "RHS Norm: " << rhsNorm << std::endl;
 
-        double initSolNorm = mechSolVec->L2Norm();
+        double initSolNorm = static_cast<double>( mechSolVec->L2Norm() );
 
         AMP::pout << "Initial Solution Norm: " << initSolNorm << std::endl;
 
         bvpOperator->residual( mechRhsVec, mechSolVec, mechResVec );
 
-        double initResidualNorm = mechResVec->L2Norm();
+        double initResidualNorm = static_cast<double>( mechResVec->L2Norm() );
 
         AMP::pout << "Initial Residual Norm: " << initResidualNorm << std::endl;
 
@@ -104,7 +104,7 @@ static void linearElasticTest( AMP::UnitTest *ut, const std::string &exeName )
 
         mlSolver->solve( mechRhsVec, mechSolVec );
 
-        double finalSolNorm = mechSolVec->L2Norm();
+        double finalSolNorm = static_cast<double>( mechSolVec->L2Norm() );
 
         AMP::pout << "Final Solution Norm: " << finalSolNorm << std::endl;
 
@@ -116,7 +116,7 @@ static void linearElasticTest( AMP::UnitTest *ut, const std::string &exeName )
 
         bvpOperator->residual( mechRhsVec, mechSolVec, mechResVec );
 
-        double finalResidualNorm = mechResVec->L2Norm();
+        double finalResidualNorm = static_cast<double>( mechResVec->L2Norm() );
 
         AMP::pout << "Final Residual Norm: " << finalResidualNorm << std::endl;
